requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
asttokens==3.0.0
attrs==24.3.0
black==24.10.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
decorator==5.1.1
executing==2.1.0
Faker==33.3.1
idna==3.10
iniconfig==2.0.0
ipython==8.31.0
jedi==0.19.2
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
packaging==24.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pytest==8.3.4
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
